Constant lighting and cube maps

+++ ViroRenderer/VROGeometrySubstrateOpenGL.h
@@ -35,7 +35,7 @@ struct VROVertexAttributeOpenGL {
     GLuint index;
     GLint size;
     GLenum type;
-    int offset;
+    uintptr_t offset;
 };
 
 struct VROVertexDescriptorOpenGL {
